@@ -2,9 +2,10 @@
   "name": "angular-froala-wysiwyg-base",
   "version": "3.0.6",
   "description": "Angular 2, 4, 5, 6 and 7 bindings for Froala WYSIWYG HTML rich text editor",
-  "main": "bundles/angular-froala-wysiwyg.umd.js",
+  "main": "dist/bundles/angular-froala-wysiwyg.umd.js",
   "typings": "index.d.ts",
   "module": "index.js",
+  "private": true,
   "scripts": {
     "lite-server": "lite-server -c demo/bs-config.json",
     "demo.serve": "npm run -s build link demo.build lite-server",
@@ -47,33 +48,33 @@
   "dependencies": {
     "font-awesome": "^4.7.0",
     "froala-editor": "^3.0.6",
-    "tslib": "^1.9.0"
+    "tslib": "^1.10.0"
   },
   "peerDependencies": {},
   "devDependencies": {
-    "lite-server": "^2.5.4",
-    "@angular-devkit/build-angular": "~0.11.0",
-    "@angular/cli": "^7.1.4",
-    "@angular/common": "^7.2.0",
-    "@angular/compiler": "^7.2.0",
-    "@angular/compiler-cli": "^7.2.0",
-    "@angular/core": "^7.2.0",
-    "@angular/forms": "^7.2.0",
-    "@angular/http": "^7.2.0",
-    "@angular/language-service": "^7.2.0",
-    "@angular/platform-browser": "^7.2.0",
-    "@angular/platform-browser-dynamic": "^7.2.0",
-    "@angular/router": "^7.2.0",
+    "@angular-devkit/build-angular": "~0.900.0-rc.6",
+    "@angular/cli": "^9.0.0-rc.6",
+    "@angular/common": "^9.0.0-rc.6",
+    "@angular/compiler": "^9.0.0-rc.6",
+    "@angular/compiler-cli": "^9.0.0-rc.6",
+    "@angular/core": "^9.0.0-rc.6",
+    "@angular/forms": "^9.0.0-rc.6",
+    "@angular/language-service": "^9.0.0-rc.6",
+    "@angular/platform-browser": "^9.0.0-rc.6",
+    "@angular/platform-browser-dynamic": "^9.0.0-rc.6",
+    "@angular/router": "^9.0.0-rc.6",
     "@types/marked": "0.0.28",
-    "@types/node": "7.0.0",
+    "@types/node": "^12.11.1",
     "@types/tapable": "^1.0.4",
     "@types/webpack": "^4.39.5",
     "classlist-polyfill": "1.0.3",
+    "codelyzer": "^5.1.2",
+    "lite-server": "^2.5.4",
     "ngm-cli": "^1.0.4",
     "rxjs": "^6.3.3",
     "rxjs-compat": "^6.3.2",
     "ts-helpers": "^1.1.1",
-    "typescript": "3.2.4",
-    "zone.js": "^0.8.26"
+    "typescript": "3.6.4",
+    "zone.js": "~0.10.2"
   }
 }
